Fix build warnings by replacing FLAnimatedImageView swizzling with proper subclass usage
- Finish work of 059f1e1a7146dadd4be0b824de4f8cff6bebb965 + https://github.com/mastodon/mastodon-ios/commit/9e486a82e68182c307ba02c0e4fab4aa95478a6b
