+++ MastodonSDK/Sources/MastodonUI/View/Content/CondensedUserView.swift
@@ -1,7 +1,6 @@
 // Copyright © 2023 Mastodon gGmbH. All rights reserved.
 
 import UIKit
-import CoreDataStack
 import MetaTextKit
 import MastodonAsset
 import MastodonLocalization
